+++ test/ecto/adapters/clickhouse/connection_test.exs
@@ -2874,14 +2874,15 @@ defmodule Ecto.Adapters.ClickHouse.ConnectionTest do
   end
 
   test "build_params/3" do
-    params = [
-      1,
-      "a",
-      true,
-      Date.utc_today(),
-      DateTime.utc_now(),
-      DateTime.utc_now() |> DateTime.truncate(:second)
-    ]
+    params =
+      [
+        1,
+        "a",
+        true,
+        ~D[2024-04-12],
+        ~U[2024-04-12 09:55:54.329788Z],
+        ~U[2024-04-12 09:55:54Z]
+      ]
 
     assert to_string(Connection.build_params(_ix = 0, _len = 0, params)) == ""
     assert to_string(Connection.build_params(_ix = 1, _len = 0, params)) == ""
@@ -2916,5 +2917,14 @@ defmodule Ecto.Adapters.ClickHouse.ConnectionTest do
 
     assert to_string(Connection.build_params(_ix = 0, _len = 6, params)) ==
              "{$0:Int64},{$1:String},{$2:Bool},{$3:Date},{$4:DateTime64},{$5:DateTime}"
+
+    assert to_string(
+             Connection.build_params(
+               _ix = 0,
+               _len = 6,
+               Enum.map(params, &Connection.mark_inline/1)
+             )
+           ) ==
+             "1,'a',true,'2024-04-12'::date,'2024-04-12 09:55:54.329788'::DateTime64(6,'Etc/UTC'),'2024-04-12 09:55:54'::DateTime('Etc/UTC')"
   end
 end